% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMacMod.R
\name{findMacMod}
\alias{findMacMod}
\title{Cluster metabolic features based on covarying abundances into modules}
\usage{
findMacMod(
  se,
  w,
  chem_tax = NULL,
  annot = NULL,
  mms = NULL,
  evaluateMOS = TRUE
)
}
\arguments{
\item{se}{SummarizedExperiment object created using MACARRoN::makeSumExp().}

\item{w}{distance matrix from function MACARRoN::makeDisMat().}

\item{chem_tax}{chemical taxonomy file with 3 columns specifying annotation, subclass and class of annotated features. 
Can be created using the decorateID.R utility of MACARRoN. 
Annotation specified with "annot" and annotation in the first column of the chemical taxonomy file must match.}

\item{annot}{a feature annotation of choice e.g. HMDB ID/Pubchem CID/Metabolite name. Default: Column 1 in annotation table.}

\item{mms}{minimum module size to be used for module identification with dynamicTreeCut::cutreeDynamic(). 
Default is cube root of number of prevalent features.}

\item{evaluateMOS}{examine measure of success for modules identified using mms, mms + 5, mms + 10, mms - 5, mms - 10}
}
\value{
mod.assn metabolic features clustered into "modules" based on covarying abundances
}
\description{
Cluster metabolic features based on covarying abundances into modules
}
\examples{
mod.assn <- findMacMod(se, w, chem_tax)

}
