% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSumExp.R
\name{makeSumExp}
\alias{makeSumExp}
\title{Create a SummarizedExperiment object}
\usage{
makeSumExp(feat_int, feat_anno, exp_meta)
}
\arguments{
\item{feat_int}{a dataframe (features x samples) containing metabolic feature intensities (abundances).}

\item{feat_anno}{a dataframe (features x annotations) containing the available feature annotations.
^^Column 1 must contain standard annotations such as HMDB ID or Pubchem CID for 
the subset of identified/annotated features.}

\item{exp_meta}{a dataframe (samples x metadata) containing sample metadata.
^^Column 1 must identify samples. It will be used for creating rownames.
^^Column 2 must identify phenotypes or conditions (categorical metadata) associated with the samples. 
Must not contain NA. Rows with no specified phenotype/condition will be removed.}
}
\value{
SummarizedExperiment object
}
\description{
Create a SummarizedExperiment object
}
\examples{
se <- makeSumExp(feat_int,feat_anno,exp_meta)

}
