% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calAVA.R
\name{calAVA}
\alias{calAVA}
\title{Calculate abundance versus anchor (AVA) of metabolic features.}
\usage{
calAVA(se, mod.assn, ptype = NULL, anchor.anno = NULL)
}
\arguments{
\item{se}{SummarizedExperiment object created using MACARRoN::makeSumExp().}

\item{mod.assn}{the output of MACARRoN::findMacMod().}

\item{ptype}{metadata of interest. Default: Column 2 of metadata table.}

\item{anchor.anno}{anchor identification. Default: column 2 of annotation dataframe.
Note: ptype must be consistent across ava, q-value and effect-size calculations.}
}
\value{
mac.ava
}
\description{
AVA of a feature is the ratio of its abundance and the most abundant metabolite in the same module 
i.e. the "anchor". Anchor is an annotated/known feature if available or just the most abundant metabolic
feature. For every feature, mean abundance in each phenotype or condition is calculated and the maximum is considered
for AVA calculation. Singletons are assigned an AVA of 1.
}
\examples{
mac.ava <- calAVA(mbx, mod.assn)

}
