% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calQval.R
\name{calQval}
\alias{calQval}
\title{Calculate q-value of differential abundance of metabolic features.}
\usage{
calQval(
  se,
  mod.assn,
  ptype = NULL,
  fix.eff = NULL,
  ran.eff = NULL,
  reference = NULL,
  output.folder = NULL,
  ncores = 1
)
}
\arguments{
\item{se}{SummarizedExperiment object created using MACARRoN::makeSumExp().}

\item{mod.assn}{the output of MACARRoN::findMacMod().}

\item{ptype}{metadata of interest. Default: Column 2 of metadata table. 
Note: ptype must be consistent across ava, q-value and effect-size calculations.}

\item{fix.eff}{fixed effects, comma delimited e.g. c("metadata1","metadata2"). Default: all columns in metadata.}

\item{ran.eff}{random effects, comma delimited. Default: NULL.}

\item{reference}{a reference level/group in each metadata, comma delimited e.g. c("metadata1,ref1","metadata2,ref2"). 
Default: alphabetically first phenotype/condition will be used as reference. 
Note: Reference must be specified for metadata with more than 2 levels.}

\item{output.folder}{the name of the output folder where all MaAsLin2 results will be written. Default: maaslin2_output}

\item{ncores}{the number of R processes to run in parallel}
}
\value{
mac.qval
}
\description{
This function uses the MaAsLin2 package for estimating q-value of differential abundance.
Multiple fixed and random effects can be specified for fitting the multiple regression model. 
Default analysis method is "LM". Can be run on multiple cores.
ptype and ref (reference group) should be the same as the one specified for effect size calculation.
}
\examples{
mac.qval <- calQval(se, mod.assn, ptype, fix.eff, ran.eff, reference, output.folder)


}
