% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorateID.R
\name{decorateID}
\alias{decorateID}
\title{Create a chemical taxonomy table for annotated metabolic features.}
\usage{
decorateID(feat_anno)
}
\arguments{
\item{feat_anno}{a dataframe (features x annotations) containing the available feature annotations.
^^Column 1 must contain standard annotations such as HMDB ID or PubChem CID for 
the subset of identified/annotated metabolic features.}
}
\value{
dataframe containing ID (HMDB or PubChem), chemical sub class and chemical class of annotated metabolic features.

chem_tax <- decorateID(feat_anno)
}
\description{
Create a chemical taxonomy table for annotated metabolic features.
}
