% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACARRoN.R
\name{MACARRoN}
\alias{MACARRoN}
\title{MACARRoN}
\usage{
MACARRoN(
  input_abundances,
  input_annotations,
  input_metadata,
  input_taxonomy,
  output = "MACARRoN_output",
  metadata_variable = NULL,
  min_prevalence = 0.7,
  execution_mode = "serial",
  standard_identifier = NULL,
  anchor_annotation = NULL,
  min_module_size = NULL,
  fixed_effects = NULL,
  random_effects = NULL,
  reference = NULL,
  cores = 1,
  plot_heatmap = TRUE,
  plot_scatter = FALSE,
  heatmap_first_n = 50
)
}
\arguments{
\item{input_abundances}{a comma-delimited file or dataframe (features x samples) containing metabolic feature intensities (abundances).}

\item{input_annotations}{a comma-delimited file or dataframe (features x annotations) containing available feature annotations.}

\item{input_metadata}{a comma-delimited file or dataframe (samples x metadata) containing sample metadata.}

\item{input_taxonomy}{a comma-delimited file or dataframe containing the chemical class and subclass information of annotated features.}

\item{output}{name of the folder where MACARRoN output files will be written. Default: "MACARRoN_output".}

\item{metadata_variable}{the header of the column in the metadata file that identifies the main phenotypes/conditions in the study. Default: Column 2 of metadata file.}

\item{min_prevalence}{prevalence threshold (percentage). Default = 0.7.}

\item{execution_mode}{BiocParallel execution mode. Options: "serial" or "multi" Default = "serial".}

\item{standard_identifier}{HMDB ID or PubChem CID. Default: Column 2 of annotation file/Column 1 of annotation dataframe.}

\item{anchor_annotation}{Metabolite name. Default: Column 3 of annotation file/Column 2 of annotation dataframe.}

\item{min_module_size}{Integer that defines the size of the smallest covariance module. Default: Cube root of number of prevalent metabolic features.}

\item{fixed_effects}{Covariates for linear modeling with MaAsLin2. Default: All columns of metadata dataframe.}

\item{random_effects}{Random effects for linear modeling with MaAsLin2. Default: NULL.}

\item{reference}{Reference category (factor) in categorical metadata covariates containing three or more levels. Must be provided as a string of 'covariate,reference' semi-colon delimited for multiple covariates.}

\item{cores}{MaAsLin2 option-The number of R processes to be run in parallel.}

\item{plot_heatmap}{MaAslin2 option-Generate a heatmap for the significant associations. Default: TRUE}

\item{plot_scatter}{MaAslin2 option-Generate scatter plots for the significant associations. Default: FALSE}

\item{heatmap_first_n}{MaAslin2 option-Generate heatmap for top n significant associations. Default: 50}
}
\value{
data.frame containing metabolic features listed according to their priority (potential bioactivity) in a phenotype of interest.
}
\description{
MACARRoN
}
\examples{
prism_abundances = system.file("extdata", "demo_abundances.csv", package="Macarron")
prism_annotations = system.file("extdata", "demo_annotations.csv", package="Macarron")
prism_metadata = system.file("extdata", "demo_metadata.csv", package="Macarron")
met_taxonomy = system.file("extdata", "demo_taxonomy.csv", package="Macarron")
mets.prioritized <- Macarron::MACARRoN(input_abundances = prism_abundances,
                                       input_annotations = prism_annotations,
                                       input_metadata = prism_metadata,
                                       input_taxonomy = met_taxonomy)

}
