% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDisMat.R
\name{makeDisMat}
\alias{makeDisMat}
\title{Create a biweight midcorrelation (WGCNA::bicor()) based distance matrix.}
\usage{
makeDisMat(
  se,
  ptype = NULL,
  preval = 0.7,
  execution_mode = "serial",
  optimize.for = c("runtime", "memory")
)
}
\arguments{
\item{se}{SummarizedExperiment object created using MACARRoN::makeSumExp().}

\item{ptype}{metadata (phenotype/condition) to be used to evaluate prevalence of features. Default = Column 2 of metadata table.}

\item{preval}{prevalence threshold (percentage). Default = 0.7.}

\item{execution_mode}{serial or multi processing with BiocParallel. Default: "serial" (recommended for laptops). 
"multi" may be used when running MACARRoN on a cluster. 

Features present (i.e. not NA) in "preval" of samples in each category of a "ptype" will be considered 
e.g. if preval is 0.7 and ptype has 2 categories A and B, union of (i) features present in at least 70% of A samples
and (ii) features present in at least 70% of B samples, will be considered for distance matrix generation. 
Correlation between feature abundances are is calculated using WGCNA::bicor().}
}
\value{
w distance matrix where distance = 1-bicor^3
}
\description{
Create a biweight midcorrelation (WGCNA::bicor()) based distance matrix.
}
\examples{
se <- makeSumExp(feat_int, feat_anno, exp_meta)
w <- makeDisMat(se)

}
