% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritize.R
\name{prioritize}
\alias{prioritize}
\title{Rank metabolic features and prioritize based on predicted bioactivity.}
\usage{
prioritize(se, mod.assn, mac.ava, mac.qval, mac.es)
}
\arguments{
\item{se}{SummarizedExperiment object created using MACARRoN::makeSumExp()}

\item{mod.assn}{the output of MACARRoN::findMacMod()}

\item{mac.ava}{the output of MACARRoN::calAVA()}

\item{mac.qval}{the output of MACARRoN::calQval()}

\item{mac.es}{the output of MACARRoN::calES()}
}
\value{
mac.result - metabolic features listed according to priority
}
\description{
Metabolic features are ranked based on AVA, and q-value and effect size of
differential abundance. The harmonic mean of these three ranks is calculated and used as 
the meta-rank to prioritize potentially bioactive features in a phenotype (or condition). 
Top-ranked features have good relative abundance, and are significantly perturbed 
in the specified environment/phenotype.
}
\examples{
mac.result <- prioritize(se, mod.assn, mac.ava, mac.qval, mac.es)

}
